Move class RssFeedService to new XmlGenerator namespace
