+++ packages/framework/tests/Feature/Services/RssFeedServiceTest.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature\Services;
 
 use Hyde\Facades\Features;
-use Hyde\Framework\Services\RssFeedService;
+use Hyde\Framework\Features\XmlGenerators\RssFeedService;
 use Hyde\Hyde;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Framework\Services\RssFeedService
+ * @covers \Hyde\Framework\Features\XmlGenerators\RssFeedService
  */
 class RssFeedServiceTest extends TestCase
 {
@@ -38,7 +38,7 @@ class RssFeedServiceTest extends TestCase
         config(['site.name' => 'Test Blog']);
         config(['site.url' => 'https://example.com']);
 
-        $service = new RssFeedService();
+        $service = new \Hyde\Framework\Features\XmlGenerators\RssFeedService();
         $this->assertObjectHasAttribute('title', $service->feed->channel);
         $this->assertObjectHasAttribute('link', $service->feed->channel);
         $this->assertObjectHasAttribute('description', $service->feed->channel);
@@ -52,7 +52,7 @@ class RssFeedServiceTest extends TestCase
     {
         config(['site.url' => 'https://example.com']);
 
-        $service = new RssFeedService();
+        $service = new \Hyde\Framework\Features\XmlGenerators\RssFeedService();
         $this->assertObjectHasAttribute('link', $service->feed->channel);
         $this->assertEquals('https://example.com', $service->feed->channel->link);
         $this->assertEquals('https://example.com/feed.xml',
@@ -69,7 +69,7 @@ class RssFeedServiceTest extends TestCase
         config(['site.url' => 'https://blog.foo.com/bar']);
         config(['hyde.rss_description' => 'Foo is a web log about stuff']);
 
-        $service = new RssFeedService();
+        $service = new \Hyde\Framework\Features\XmlGenerators\RssFeedService();
         $this->assertEquals('Foo', $service->feed->channel->title);
         $this->assertEquals('https://blog.foo.com/bar', $service->feed->channel->link);
         $this->assertEquals('Foo is a web log about stuff', $service->feed->channel->description);
@@ -120,13 +120,13 @@ class RssFeedServiceTest extends TestCase
 
     public function test_get_xml_method_returns_xml_string()
     {
-        $service = new RssFeedService();
+        $service = new \Hyde\Framework\Features\XmlGenerators\RssFeedService();
         $this->assertStringStartsWith('<?xml version="1.0" encoding="UTF-8"?>', ($service->getXML()));
     }
 
     public function test_generate_feed_helper_returns_xml_string()
     {
-        $this->assertStringStartsWith('<?xml version="1.0" encoding="UTF-8"?>', (RssFeedService::generateFeed()));
+        $this->assertStringStartsWith('<?xml version="1.0" encoding="UTF-8"?>', (\Hyde\Framework\Features\XmlGenerators\RssFeedService::generateFeed()));
     }
 
     public function test_can_generate_feed_helper_returns_true_if_hyde_has_base_url()
